% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areaCov.R
\name{areaCov}
\alias{areaCov}
\title{Coverage plot of annotations around TAD boundaries}
\usage{
areaCov(data.gr, annot.strand = TRUE, bin.width = NULL, norm = FALSE)
}
\arguments{
\item{data.gr}{Output of TADarea function.}

\item{annot.strand}{If TRUE (default): separate coverage according to their strands.}

\item{bin.width}{Size of the sliding window to calculate the mean coverage. By default the size in 10 time smaller than window.size.}

\item{norm}{Normalized coverage according to the total width of the considered strand. Default is FALSE.}
}
\value{
Return a ggplot graph
}
\description{
Graph of the coverage of the genomic annotations around TAD boundaries (real distance).
}
\details{
This function take the output of TADarea function.
If the cumulative size is significantly different between strands, it can be useful to normalized the depth between strand (using norm = TRUE).
}
\examples{
# create GRange with 1 TAD
tad.gr <- dataframes2grange(
  data.frame(chr = 1, start = c(10e3), end = c(50e3)),
  data.frame(chr = "1", size = 60e3)
)

# create annotations of 12 genes every 2kb with width = 10kb on both strand:
annot.gr <- dataframes2grange(
  data.frame(
    chr = rep("1", 12),
    start = seq(20e3, 42e3, 2e3),
    end = seq(20e3, 42e3, 2e3) + 10e3,
    strand = c(rep("+", 6), rep("-", 6))
  ),
  data.frame(chr = "1", size = 60e3),
  strand.col = 4
)
# Vizualisation
TADplot(tad.gr = tad.gr, start = 5e3, stop = 55e3, chr = 1, annot.gr = annot.gr)


# Distribution analysis surrounding the start of the TAD
data.gr <- TADarea(
  tad.gr = tad.gr,
  annot.gr = annot.gr,
  window.size = 60e3,
  tad.boundary = "start"
)

# plot coverage
areaCov(data.gr = data.gr, annot.strand = FALSE)
areaCov(data.gr = data.gr, annot.strand = TRUE)

}
