% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areaHist.R
\name{areaHist}
\alias{areaHist}
\title{Histogram plot of annotations around TAD boundaries}
\usage{
areaHist(
  data.gr,
  annot.border = "center",
  annot.strand = TRUE,
  bin.width = NULL
)
}
\arguments{
\item{data.gr}{Output of TADarea function.}

\item{annot.border}{Type of feature to analyzed. Start, end or center of each annotations from annot.gr files. It is possible to use "closest" to take the closest border (from TAD boundary) independently to the start or end (or independently to the strands).}

\item{annot.strand}{Default is FALSE to plot the distribution as histogram. If TRUE, distributions are separated according to their strands and are display with lines.}

\item{bin.width}{Size of the bin to count the number of annotations features. If NULL the bin size is 10 times smaller than window.size. It is recommended to use the same size as the HiC matrix.}
}
\value{
Return a ggplot graph
}
\description{
Graph of the distribution of the genomic annotations around TAD boundaries (real distance).
Measures the distances between TAD boundaries and the annotation borders (start, end, closest or the center) and plot the distribution as an histogram.
This function take the output of TADarea function.
annot.border="closest", use the closest border (start or stop) to measure the distances from the TAD boundary independently to his strand. Therefore all annotations that overlap a TAD boundary have a distance of 0.
}
\details{
Several features distances can be count. As an example (see examples for visualizations) of 2 genes which start at 20Kb of a TAD boundary (width of the 2 genes are 10Kb and are on the reverse an forward strand respectively).
-If annot.border = "start", the distances from the TAD border are 20Kb,
-If annot.border = "end", the distances are 10 and 30Kb respectively,
-If annot.border = "center", the distances are 15Kb and 25Kb respectively,
-If annot.border = "closest", the distances are 10Kb and 20Kb respectively.
}
\examples{
# create GRange with 2 annotations (genes with same size and start but different strand)
annot.gr <- tad.gr <- dataframes2grange(
  data.frame(chr = 1, start = c(20e3, 30e3), end = c(30e3, 40e3), strand = c("-", "+")),
  data.frame(chr = "1", size = 60e3),
  strand.col = 4
)

# create GRange with 1 TAD
tad.gr <- dataframes2grange(
  data.frame(chr = 1, start = c(10e3), end = c(50e3)),
  data.frame(chr = "1", size = 60e3)
)

annot.gr
tad.gr

# Vizualisation
TADplot(tad.gr = tad.gr, start = 5e3, stop = 55e3, chr = 1, annot.gr = annot.gr)

# Distribution analysis surrounding the start of the TAD
data.gr <- TADarea(
  tad.gr = tad.gr,
  annot.gr = annot.gr,
  window.size = 50e3,
  tad.boundary = "start"
)

# Visualized differences according to annotation features:
for (features in c("start", "end", "center", "closest")) {
  output <- areaHist(
    data.gr = data.gr,
    annot.strand = FALSE,
    bin.width = 1e3,
    annot.border = features
  )
  print(output)
}
}
