% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgCorr.R
\name{bgCorr}
\alias{bgCorr}
\title{Correlation between bedgraph scores}
\usage{
bgCorr(bedgraph.lst, method = "pearson", rm_chr = "X")
}
\arguments{
\item{bedgraph.lst}{list of \code{data.frame}, \code{GRanges} or full path of the bedgraph files (data frame without header and 4 columns tab separated) containing a score for each bin.}

\item{method}{a character string indicating which correlation coefficient is to be computed. One of "pearson" (default), "kendall", or "spearman": can be abbreviated.}

\item{rm_chr}{chromosome to filter, default = "X".}
}
\value{
matrix with correlation values
}
\description{
Computes the correlations between bin scores (e.g. between insulation scores...) from 2 or more individuals.
}
\examples{
bgCorr(list(ind1 = IS_1_10kb.bedgraph, ind2 = IS_2_10kb.bedgraph))

}
