% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary2domain.R
\name{boundary2domain}
\alias{boundary2domain}
\title{Convert boundary to domain}
\usage{
boundary2domain(boundaries)
}
\arguments{
\item{boundaries}{\code{dataframe} (chromosome, start and end) or \code{GRanges} with the boundaries.}
}
\value{
\code{GRanges} object with domains
}
\description{
From datas with boundaries stored by line (chromosome, start and end), \code{boundary2domain()} return \code{GRanges} with domains.
}
\details{
Start and end of domains are the middle of boundaries.
}
\examples{
#create GRanges with chr sizes:
boundaries.gr = dataframes2grange(tad_HCT116_5kb.bed, human_chromsize)

domains.gr = boundary2domain(boundaries.gr)
domains.gr

}
