% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chromsize}
\alias{chromsize}
\title{Size of cow chromosomes 25 and 26.}
\format{
\subsection{\code{chromsize}}{

A data frame containing the size of the two bovine chromosomes
\describe{
\item{chr}{chromosome name}
\item{size}{sine in base pair}
}
}
}
\usage{
chromsize
}
\description{
Size of cow chromosomes 25 and 26.
}
\keyword{datasets}
