% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cool2matrix.R
\name{cool2matrix}
\alias{cool2matrix}
\title{Import matrix from cool or mcool file}
\usage{
cool2matrix(
  cool.path,
  chr,
  bin.width = NA,
  balance = FALSE,
  balancing_name = "weight"
)
}
\arguments{
\item{cool.path}{The full path of the cool or mcool file.}

\item{chr}{The selected chromosome.}

\item{bin.width}{Bin width in base pair (i.e resolution) for mcool file. Default is \code{NA} for cool file.}

\item{balance}{Logical. Weather or not to use balanced counts instead of raw counts. Default = \code{FALSE}.}

\item{balancing_name}{Character that must correspond to the name given to the normalization method used. The most common names (for HiC normalisation) are "weight" (default), "KR", "VC".}
}
\value{
A \code{dgCMatrix} object: upper triangular and sparse Matrix
}
\description{
From cool or mcool files, \code{cool2matrix} import the interaction matrix for one chromosome as a \code{dgCMatrix} (upper triangular and sparse Matrix).
If balance = \code{TRUE}, \code{cool2matrix} returns the normalized counts.
}
\details{
The cool file format is an efficient storage format for high resolution genomic interaction matrices, developed and maintained by the Mirny lab (https://github.com/open2c/cooler)
\code{cool2matrix} use the indexes provided by the cooler data model to extract the intra-chromosomal counts for a given chromosome.
As cool files store genomic interactions for only one resolution, \code{bin.width} must be set to \code{NA}.
While, as mcool files store genomic interactions for multiples resolutions, the chosen resolution must be set with the \code{bin.width} parameter.
}
