% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coolFetch.R
\name{coolFetch}
\alias{coolFetch}
\title{Fetch a chromosome of a mcool file}
\usage{
coolFetch(path, resolution, chrom)
}
\arguments{
\item{path}{the full path of the mcool file}

\item{resolution}{the desired resolution}

\item{chrom}{the selected chromosome}
}
\value{
a sparse matrix
}
\description{
Using the Indexes provided byt the cooler data model :
https://cooler.readthedocs.io/en/latest/schema.html,
extract the intra-chromosomal counts for a given chromosome
}
