% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframes2grange.R
\name{dataframes2grange}
\alias{dataframes2grange}
\title{Create GRange with sequence lengths}
\usage{
dataframes2grange(
  annotation.table,
  chromsize,
  chr.col = 1,
  start.col = 2,
  end.col = 3,
  strand.col = NULL,
  name.col = NULL,
  metadata.mcols = NULL
)
}
\arguments{
\item{annotation.table}{Data frame with genomic annotations (at least 3 columns: chromosome, start and stop).}

\item{chromsize}{Data frame with chromosome names and lengths. Chromosomes are also filtered according to this list.}

\item{chr.col}{Column number for chromosome names, default=1.}

\item{start.col}{Column number for start, default=2.}

\item{end.col}{Column number for end, default=3.}

\item{strand.col}{Column number for strands, default=NULL to set strands as "*" (instead of "+" or "-").}

\item{name.col}{Column number for annotation names, default=NULL to create names with chr.names and start.}

\item{metadata.mcols}{Column number(s) for metadata, default=NULL (ie no metadata).}
}
\value{
GRange file that contains chromosome size.
}
\description{
This function create a GRange object from the first dataframe and add the size of the chromosomes from the second dataframe.
A name of each line is added to the GRange object.
}
