% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domainDens.R
\name{domainDens}
\alias{domainDens}
\title{Distribution of annotation density within domains}
\usage{
domainDens(
  domain.gr,
  annot.gr,
  domain.col = NULL,
  annot.col = NULL,
  bin.width = 50000,
  output = "plot"
)
}
\arguments{
\item{domain.gr}{GRange file with domains (TADs, compartments...).}

\item{annot.gr}{GRange file with genomic annotations (genes, repeat elements...).}

\item{domain.col}{Column number (metadata columns) used to separate domains (e.g. column to differentiate between compartments A and B). By default, domain.col = NULL.}

\item{annot.col}{Column number (metadata columns) used to separate annotations. By default, annot.col = NULL. Use "strand" in order to separate annotations according to their strands.}

\item{bin.width}{Size of the bin in bp. This should match the size of the bins used to determine the domains.}

\item{output}{Default is "plot" to return a ggplot. Use "data" to return the datas used to produce the plot.}
}
\value{
Return a ggplot graphs
}
\description{
Distribution of annotations density (e.g. genes) within domains (e.g. TADs, compartments...).
This function calculates the density per bin of annotations and measures the relative position of each bin within each domain
}
\details{
As an exemple, this function take all TAD domains and calculate the density of annotations .
For example:
1- calculates the density of genes per bin of 10kb,
2- calculate the relative positions of each bin within his corresponding TAD,
3- plot the smoothed bin density (and smoothed zscore bin density) according to relative positions.
}
\examples{
#to do

}
