% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domainHist.R
\name{domainHist}
\alias{domainHist}
\title{Histogram of annotation distribution within domains}
\usage{
domainHist(
  domain.gr,
  annot.gr,
  annot.boundary = "start",
  ifoverlap = "remove",
  annot.strand = FALSE,
  bin.width = 5,
  output = "plot"
)
}
\arguments{
\item{domain.gr}{\code{GRanges} with domains (TADs, compartments...).}

\item{annot.gr}{\code{GRanges} with genomic annotations (genes, repeat elements...).}

\item{annot.boundary}{Type of feature to analyzed. \code{"Start"}, \code{"end"} or \code{"center"} of each \code{annot.gr}. It is possible to use \code{"closest"} to take the closest border (start or end) independently to the strands.}

\item{ifoverlap}{In case of annotation overlap a TAD boundary, few options are available to measure the \code{annot.boundary} positions:
\itemize{
\item \code{"remove"} to remove all \code{annot.gr} that overlaps a TAD boundary,
\item \code{"real"} to take the position according to the TAD where the \code{annot.boundary} is located,
\item \code{"best"} to take the position according to the TAD where the \code{annot.gr} has the best overlay.
}}

\item{annot.strand}{Default is \code{FALSE} to plot the distribution as histogram. If \code{TRUE}, distributions are separated according to their strands and are displayed with lines.}

\item{bin.width}{Size of the bin in percent to count the number of annotations features, default is 5\%. if \code{ifoverlap = "best"}: the real bin distances (measured outside of domains) are equal to \code{bin.width * 1000}. Therefore default real bin size is 5kb.}

\item{output}{Default is \code{"plot"} to return a \code{ggplot}. Use \code{"data"} to return the datas used to produce the plot.}
}
\value{
\code{ggplot} object.
}
\description{
Plot of the distribution of the genomic annotations within the domains (relative distance)
Measures the relative distances of annotation border (start, end) within the domain and plot the distribution as an histogram.
}
\details{
\code{domainHist()} measures the relative positions of annotation features (start, end, closest or the center of annotations, more details in \code{areahist()}) within domains and plot the distribution as an histogram.
It is possible that some annotations overlap a boundary, in this cases few option are available to get the relative positions (see \code{ifoverlap} parameter and the example for better understanding):
* remove those annotations,
* uses the domain in which the feature (\code{"start"}, \code{"end"} or \code{"center"} of the annotation) is located,
* uses the domain in which the annotation has the best overlap.
Therefore, in some cases, the \code{"start"} of an \code{annot.gr} can be located before the domain (i.e before the domain with the best overlap). In that case, the distance between the \code{"start"} of \code{annot.boundary} and the domain is represented with the real distance (in base pair).
}
\examples{
#Create 20 genes
annot.gr <- dataframes2grange(
  data.frame(chr = 1, start = seq(190.5e3, 209.5e3, 1e3), end = 290e3, strand = "+"),
  data.frame(chr = "1", size = 400e3),
  strand.col = 4
)
#Create 2 TADs
tad.gr <- dataframes2grange(
  data.frame(chr = 1, start = c(100e3, 200e3), end = c(200e3, 300e3)),
  data.frame(chr = "1", size = 400e3)
)

TADplot(tad.gr = tad.gr, annot.gr = annot.gr, start = 150e3, stop = 300e3, chr = 1)

#see distribution of gene starts according to the "ifoverlap" parameters:
domainHist(domain.gr = tad.gr, annot.gr = annot.gr, ifoverlap = "real")
domainHist(domain.gr = tad.gr, annot.gr = annot.gr, ifoverlap = "remove")
domainHist(domain.gr = tad.gr, annot.gr = annot.gr, ifoverlap = "best")

}
