% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneTADtopo.R
\name{geneTADtopo}
\alias{geneTADtopo}
\title{Genes topology according to TAD}
\usage{
geneTADtopo(
  domain.gr,
  annot.gr,
  ifoverlap = "best",
  expression.data.frame = NULL
)
}
\arguments{
\item{domain.gr}{\code{GRanges} with domains (TADs, compartments...).}

\item{annot.gr}{\code{GRanges} with genomic annotations.}

\item{ifoverlap}{In case of annotation overlap a TAD boundary, few options are available to measure the \code{annot.boundary} positions:
\itemize{
\item \code{"remove"} to remove all \code{annot.gr} that overlaps a TAD boundary,
\item \code{"real"} to take the position according to the TAD where the \code{annot.boundary} is located,
\item \code{"best"} to take the position according to the TAD where the \code{annot.gr} has the best overlay.
}}

\item{expression.data.frame}{\code{dataframe} with expression data (raw counts...).
The first two columns should be used to identify the genes. The 1st one must use the same gene id than \code{names(annot.gr)} (i.e unique ID). The 2nd can be used with usual gene names.
Others columns give expression count, one column per sample/experiment.}
}
\value{
Return a S3 class object with 3 data frames.
}
\description{
From the annotation of TADs and genes, this function return 3 data frames:
1- for each TAD: numbers of genes, their strands and IDs (and the expression data if it is provided).
It also return some statistics:
2- the number of TAD according to the number of genes,
3- the number of TAD according to the number of genes and strand.
}
\details{
Return a list of 3 dataframes.
}
\examples{
# not run

# get gene annotation
# txdb <- makeTxDbFromBiomart(biomart = "ensembl", dataset = "btaurus_gene_ensembl")
# genomic.gr = genes(txdb)
#
# create expression data set with random raw count
# expression.data.frame = data.frame(ID = names(genes.gr), Name = names(genes.gr),
#   raw_count = sample(1:100, length(genes.gr), replace = TRUE))

# output <- geneTADtopo(dataframes2grange(tad_1_10kb.bed, chromsize),
#   genomic.gr, expression.data.frame = expression.data.frame)

}
