% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human_chromsize}
\alias{human_chromsize}
\title{Size of Human autosomes.}
\format{
\subsection{\code{human_chromsize}}{

A data frame containing the size in base pair of human autosomes
\describe{
\item{chr}{chromosome name}
\item{size}{chromosome size in base pair}
}
}
}
\usage{
human_chromsize
}
\description{
Size of Human autosomes.
}
\keyword{datasets}
