% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IS_HCT116_chr19_5kb.bedgraph}
\alias{IS_HCT116_chr19_5kb.bedgraph}
\title{Insulation score of Human chromosome 19.}
\format{
\subsection{\code{IS_HCT116_chr19_5kb.bedgraph}}{

A data frame containing insulation score.
For chromosome 19 from Human HCT116 cells with a bin width of 5kb.
Downloaded from the 4DN portal with accession number 4DNESNSTBMBY.
\describe{
\item{chr}{chromosome name}
\item{start}{bin start}
\item{end}{bin end}
\item{IS}{insulation score}
}
}
}
\usage{
IS_HCT116_chr19_5kb.bedgraph
}
\description{
Insulation score of Human chromosome 19.
}
\keyword{datasets}
