% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mat_HCT116_chr19_50kb}
\alias{mat_HCT116_chr19_50kb}
\title{HiC matrix of Human HCT116 cell.}
\format{
\subsection{\code{mat_HCT116_chr19_50kb}}{

A dgCMatrix object with 1173 rows and columns containing the raw counts (number of interactions).
For chromosome 19 from Human HCT116 cells with a bin width of 50kb.
Downloaded from the 4DN portal with accession number 4DNESNSTBMBY.
}
}
\usage{
mat_HCT116_chr19_50kb
}
\description{
HiC matrix of Human HCT116 cell.
}
\keyword{datasets}
