% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matExpDist.R
\name{matExpDist}
\alias{matExpDist}
\title{interactions observed versus genetic distances}
\usage{
matExpDist(
  matrix.lst,
  bin.width,
  colors.lst = c("#66C2A5", "#FC8D62", "#8DA0CB", "#E78AC3", "#A6D854", "#FFD92F",
    "#E5C494", "#B3B3B3")
)
}
\arguments{
\item{matrix.lst}{List of \code{dgCMatrix} or \code{matrix} object for only one chromosome.}

\item{bin.width}{Bin width of the matrices in base pair.}

\item{colors.lst}{Set of 8 colors used for plot.}
}
\value{
ggplot
}
\description{
This function calculates the average number of interactions (i.e. expected number of interaction) as a function of the distance between bins.
}
\examples{
matExpDist(matrix.lst = list(HCT116 = mat_HCT116_chr19_50kb), bin.width = 50e3)


}
