% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MATfeatures.R
\name{MATfeatures}
\alias{MATfeatures}
\title{Pileup of matrices surrounding genomic features}
\usage{
MATfeatures(
  matrix,
  bin.width,
  annot.gr,
  chr,
  annot.boundary = "start",
  window.size = NULL,
  output = "matrix"
)
}
\arguments{
\item{matrix}{\code{dgCMatrix} or \code{matrix} object for only one chromosome.}

\item{bin.width}{Bin width of the matrix in base pair.}

\item{annot.gr}{\code{GRanges} with genomic annotations.}

\item{chr}{The selected chromosome used to filter \code{annot.gr}.}

\item{annot.boundary}{Type of feature to analyzed. \code{"Start"}, \code{"end"} or \code{"center"} of each \code{annot.gr}.}

\item{window.size}{Window to analyze the matrix on each side of the features. By default the window is 40 times the \code{bin.width}.}

\item{output}{Default is \code{"matrix"} to return the \code{matrix} (i.e sum of matrices). Use \code{"plot"} to return a \code{ggplot} of the observed / expected of the pileup of the matrices, i.e. ObsExp(sum of matrices).}
}
\value{
A \code{dgCMatrix} object: upper triangular and sparse Matrix
}
\description{
This function allows the quantification of patterns (of the matrix) surrounding genomic features.
For example this function allow to visualize the patterns of the matrices around TADs boundary.
}
\details{
From genomic features (e.g. start of TADs):
\itemize{
\item export the matrix +/- \code{window.size} of each genomic features,
\item pileup the matrices (i.e. sum all the matrices),
\item return the pileup matrix or plot the observed/expected of the pileup matrix.
}
}
\examples{
boundaries.gr = dataframes2grange(tad_HCT116_5kb.bed, human_chromsize)
domains.gr = boundary2domain(boundaries.gr)
MATfeatures(matrix = mat_HCT116_chr19_50kb, bin.width = 50e3, annot.gr = domains.gr, chr = "chr19",
            annot.boundary = "start", window.size = 1e6, output = "plot")

}
