% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matObsExp.R
\name{matObsExp}
\alias{matObsExp}
\title{observed / expected matrix}
\usage{
matObsExp(matrix, matrix.colname = T, matrix.rowname = T, matrix.sep = "\\t")
}
\arguments{
\item{matrix}{R object (data frame or matrix) or the matrix file path (in matrix format) for only one chromosome. The path can be gzip file (ending by .gz)}

\item{matrix.colname}{logical. Does your matrix file (ie path) have column names (ie header)? Default = TRUE}

\item{matrix.rowname}{logical. Does your matrix file (ie path) have row names? Default = TRUE}

\item{matrix.sep}{the field separator character. Values on each line of the matrix file are separated by this character. Default = '\\t' (ie tabulation).}
}
\value{
matrix.
}
\description{
For each bin of the matrix (interactions) this function return a ratio: observed / expected.
The input file path must be in a matrix format (ie. array with as many rows and columns than the number of bin).
The output can be plot with MATplot function with log2=T
}
\details{
The expected interaction corresponds to the average of the interactions according to the bin distances.
}
