% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matPCA.R
\name{matPCA}
\alias{matPCA}
\title{perform principal component analysis f}
\usage{
matPCA(matrix, bin.width, input = "OE", seqname = "1", nb_PC = 1)
}
\arguments{
\item{matrix}{\code{Matrix} or \code{matrix} object.}

\item{bin.width}{Bin width of the matrix in base pair.}

\item{seqname}{chromosome names as character, default = "1".}

\item{nb_PC}{number of principal component to compute, default = 1.}

\item{intput}{input matrix: "Obs" for observed count or "OE" (default) for Observed / Expected counts.}
}
\value{
\code{GRanges} bedgraph with PC score.
}
\description{
perform principal component analysis f
}
\examples{
matPCA(matrix = mat_HCT116_chr19_50kb,
    bin.width = 50e3,
    input = "Obs",
    seqname = "chr1")

}
