% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MATplot.R
\name{MATplot}
\alias{MATplot}
\title{Plot TADs for one individual}
\usage{
MATplot(
  matrix,
  start,
  stop,
  bin.width,
  scale.colors = "H",
  matrix.colname = T,
  matrix.rowname = F,
  matrix.sep = "\\t",
  matrix.diag = T,
  log2 = T,
  tad.upper.tri = NULL,
  tad.lower.tri = NULL,
  loop.bedpe = NULL,
  annotations.color = "red",
  tad.upper.line = NULL,
  tad.lower.line = NULL,
  line.colors = c("red", "bleu"),
  tad.chr = NULL
)
}
\arguments{
\item{matrix}{matrix object (data frame or matrix) or matrix path at matrix format (dense) for only one chromosome. The path can be gzip (ending by .gz)}

\item{start}{start in bp of the area of interest.}

\item{stop}{end in bp of the area of interest.}

\item{bin.width}{bin width of the matrix in bp.}

\item{scale.colors}{A character string indicating the color map option to use. Eight options are available (see viridis package), default is "H":
"magma" (or "A")
"inferno" (or "B")
"plasma" (or "C")
"viridis" (or "D")
"cividis" (or "E")
"rocket" (or "F")
"mako" (or "G")
"turbo" (or "H")}

\item{matrix.colname}{logical. Does your matrix file (ie path) have column names (ie header)? Default = TRUE}

\item{matrix.rowname}{logical. Does your matrix file (ie path) have row names? Default = FALSE}

\item{matrix.sep}{the field separator character. Values on each line of the matrix file are separated by this character. Default = '\\t' (ie tabulation).}

\item{matrix.diag}{logical. Weather or not to plot diagonal of the matrix. Default = TRUE}

\item{log2}{logical. Weather or not to plot the log2 of the matrix values.}

\item{tad.upper.tri}{bed files path, data frame or grange object with the TAD to plot as triangle in the upper part of the matrix. Default is NULL}

\item{tad.lower.tri}{bed files path, data frame or grange object with the TAD to plot as triangle in the upper part of the matrix. Default is NULL}

\item{annotations.color}{color for loop and tri annotations. Default is "red".}

\item{tad.upper.line}{bed files path, data frame or grange object with the TAD to plot as line in the upper parts of the matrix. Default is NULL}

\item{tad.lower.line}{bed files path, data frame or grange object with the TAD to plot as line in the lower parts of the matrix. Default is NULL}

\item{line.colors}{column number use to color tad.upper.line and tad.lower.line. Default is NULL}

\item{tad.chr}{chromosome name to filter in TAD files. Default is NULL}

\item{loop.upper.bedpe}{bedpe files path or data frame plot on both parts of the matrix. Six columns table (chr1, start1, end1, chr2, start2, end2) that gives loops between 2 areas. Default is NULL}

\item{line.color}{colors for}
}
\value{
ggplot of the matrix.
}
\description{
TADplot use Gviz package to plot at least 2 tracks. The first one includes all the TADs from the specified chromosome. And the second one is a zoom in the defined area.
Other tracks can be added like gene annotation or read depth data.
}
\details{
This create a plot with at least 2 tracks from TADs annotation (tad.gr).

On the first track the alternating TADs are represented in grey and black while the interTAD regions are represented in white.
The second track is a zoom in the area of interest. This area is extend to the border of the TADs on both sides.

Another track from a bigwig file can be added like read depth sequencing (RNAseq...) as an histogram.
The bin size of the histogram is 1Kb by default, the read depth is smoothed using the median value of each bin.
If the chromosome name is different (between the bigwig and the chr parameter), it can be fix using the bigwig.chr parameter ("1" versus "chr1").

Another track with any annotations can be added.
This track can be group using factors in a specified column of the annot.gr files (metadata) otherwise the names of each annotation is used.
}
