% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{matrix_1_chr25_50kb}
\alias{matrix_1_chr25_50kb}
\title{Matrix of individual 1 and chromosome 25.}
\format{
\subsection{\code{matrix_1_chr25_50kb}}{

A dgCMatrix object with 847 rows and columns containing the raw counts (number of interactions).
For individual 1, chromosome 25 with a bin width of 50e3 bases pairs.
Only the upper matrix and non zero count are stored.
}
}
\usage{
matrix_1_chr25_50kb
}
\description{
Matrix of individual 1 and chromosome 25.
}
\keyword{datasets}
