% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDplot.r
\name{MDplot}
\alias{MDplot}
\title{MD plot between matrices}
\usage{
MDplot(matrix.lst, bin.width = NULL, centered = FALSE)
}
\arguments{
\item{matrix.lst}{list of \code{dgCMatrix} or \code{matrix} object for only one chromosome.}

\item{bin.width}{Default is NULL, but if bin size is set, real distances between interactions is used on the x axis.}

\item{centered}{Logical, default is FALSE. Whether or not center y-axis on zero.}
}
\value{
ggplot
}
\description{
Plot of log2(mat1 / mat2) between interactions values of 2 matrices according to bin distances. Note that only 1e6 interaction pairs are randomly selected to produce the plot.
}
\examples{
#create matrices with rnadom number of interactions:
nrow = nrow(mat_HCT116_chr19_50kb)
mat2 = matrix(abs(rnorm(nrow*nrow)), nrow, nrow)

#create list of matrices
matrix.lst = list(ind1 = mat_HCT116_chr19_50kb, ind2 = mat2)

MDplot(matrix.lst, bin.width = 50e3, centered = TRUE)

}
