% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mMATplot.R
\name{mMATplot}
\alias{mMATplot}
\title{Plot 2 matrix with annotations}
\usage{
mMATplot(
  matrix.upper,
  matrix.lower,
  start,
  stop,
  bin.width,
  log2 = T,
  scale.colors = "H",
  matrix.colname = T,
  matrix.rowname = F,
  matrix.sep = "\\t",
  matrix.upper.txt = NULL,
  matrix.lower.txt = NULL,
  tad.upper.tri = NULL,
  tad.lower.tri = NULL,
  loop.bedpe = NULL,
  tad.chr = NULL,
  annotations.color = "red",
  tad.upper.line = NULL,
  tad.lower.line = NULL,
  tad.line.col = NULL,
  line.colors = c("red", "blue")
)
}
\arguments{
\item{matrix.upper}{matrix object (data frame or matrix) or matrix path for only one chromosome. The path can be gzip (ending by .gz).}

\item{matrix.lower}{matrix object (data frame or matrix) or matrix path for only one chromosome. The path can be gzip (ending by .gz).}

\item{start}{start in bp of the area of interest.}

\item{stop}{end in bp of the area of interest.}

\item{bin.width}{bin width of the matrix in bp.}

\item{log2}{logical. Use the log2 of the matrix values. Default is TRUE}

\item{scale.colors}{A character string indicating the color map option to use. Eight colors palettes are available from viridis package. Another palette "OE" is made for data centered on 0 (ie log2(observed/expected) matrix). Default is "H":
"ObsExp" (or "OE")
"magma" (or "A")
"inferno" (or "B")
"plasma" (or "C")
"viridis" (or "D")
"cividis" (or "E")
"rocket" (or "F")
"mako" (or "G")
"turbo" (or "H")}

\item{matrix.colname}{logical. Does your matrix file (ie path) have column names (ie header)? Default = TRUE}

\item{matrix.rowname}{logical. Does your matrix file (ie path) have row names? Default = FALSE}

\item{matrix.sep}{the field separator character. Values on each line of the matrix file are separated by this character. Default = '\\t' (ie tabulation).}

\item{matrix.upper.txt}{text to write on the upper part of the plot}

\item{matrix.lower.txt}{text to write on the lower part of the plot}

\item{tad.upper.tri}{bed files path, data frame or grange object with the TAD to plot as triangle in the upper part of the matrix. Default is NULL}

\item{tad.lower.tri}{bed files path, data frame or grange object with the TAD to plot as triangle in the upper part of the matrix. Default is NULL}

\item{loop.bedpe}{bedpe files path or data frame to plot on both parts of the matrix. Six columns table (chr1, start1, end1, chr2, start2, end2) that gives loops between 2 areas. Default is NULL}

\item{tad.chr}{chromosome name to filter bed and bedpe files. Default is NULL}

\item{annotations.color}{color for loop and tri annotations. Default is "red".}

\item{tad.upper.line}{bed files path, data frame or grange object with the TAD to plot as line in the upper parts of the matrix. Default is NULL}

\item{tad.lower.line}{bed files path, data frame or grange object with the TAD to plot as line in the lower parts of the matrix. Default is NULL}

\item{tad.line.col}{col number of the tad.line files that contain factors used to color tad.upper.line and tad.lower.line. Default is NULL}

\item{line.colors}{colors for upper and lower lines.}
}
\value{
ggplot
}
\description{
Same as MATplot function but mMATplot allow to plot 2 matrices on the upper or lower part of the matrix.
-domains (e.g. TADs or compartments) are plot as triangles or lines on the upper or/and lower part of the matrix.
-interaction between domains/bins (loop) are plot as squares on the upper and lower part of the matrix.
}
\details{
The matrix data can be R object (dataframe, matrix or array) or the path of the files. The matrix format has as many rows as columns and this number corresponds to the number of bin of the chromosome.
For the path of the matrix file, you can specify if the files as column and row names. These are generally used to specify the bin number or the coordinates of each bin.
All domains (TADs or compartments) are bed files (3 columns) and can be R object (dataframe or GRange) or the path of the files.
For tad.lines, another column can be used to specify compartment A or B and read with the line.colors parameter and colored with line.colors parameter.
Loop are stored in bedpe files (6 columns) and can be a dataframe or the path of the file.
Chromosome of domain and loop datas can be filter using tad.chr parameter.
}
