% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca2matrix.R
\name{orca2matrix}
\alias{orca2matrix}
\title{Import Orca matrix into full chromosome matrix}
\usage{
orca2matrix(
  df_prediction.path,
  sep = "\\t",
  mpos,
  scale,
  chromsize,
  output = "OE",
  df_normmats.path = NULL,
  model = 3.2e+07
)
}
\arguments{
\item{df_prediction.path}{The full path of the Orca matrix. Data frame without header which contains orca predictions, i.e. log(Obseved/Expected).}

\item{sep}{separator between matrix field (df_prediction.path or df_normmats.path), default is tabulation.}

\item{mpos}{The coordinate to zoom into for multiscale prediction. By default mpos is the center of the 32Mb sequence provided (i.e. start + 16Mb).}

\item{scale}{Size of the scale (i.e. window) of the predictions in base pair (i.e. 32Mb, 16Mb, 8Mb, 4Mb, 2Mb and 1Mb).}

\item{chromsize}{Chromosome size in base pair of the related chromosome.}

\item{output}{Character: Default is "OE" to return observed / expected counts. "Obs" to return observed counts.}

\item{df_normmats.path}{Optional. Background distance-based expected balanced contact matrices. Only needed to return observed counts.}

\item{model}{orca prediction model (32e6 or 1e6). Default is 32e6.}
}
\value{
A \code{Matrix} class object: upper triangular and sparse Matrix
}
\description{
This function create an empty matrix of the full chromosome which is then filled with Orca matrix (i.e. 250x250 bins).
}
\details{
From the matrices produced by orca_predict.genomepredict function.
}
