% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PC1_1_50kb.bedgraph}
\alias{PC1_1_50kb.bedgraph}
\title{First principal component score of individual 1 and chromosome 25/26.}
\format{
\subsection{\code{PC1_1_50kb.bedgraph}}{

A data frame containing PC1 score calculated by dcHiC on 50kb matrix.
For individual 1, chromosome 25 and 26.
\describe{
\item{chr}{chromosome name}
\item{start}{bin start}
\item{end}{bin end}
\item{PC1}{principal component score}
}
}
}
\usage{
PC1_1_50kb.bedgraph
}
\description{
First principal component score of individual 1 and chromosome 25/26.
}
\keyword{datasets}
