% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PC1_250kb.gr}
\alias{PC1_250kb.gr}
\title{First principal of Human chromosomes.}
\format{
\subsection{\code{PC1_1_50kb.bedgraph}}{

A GRanges containing PC1 score from Human HCT116 cells with a bin width of 250kb.
Downloaded from the 4DN portal with accession number 4DNESNSTBMBY.
}
}
\usage{
PC1_250kb.gr
}
\description{
First principal of Human chromosomes.
}
\keyword{datasets}
