% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PC1calling.R
\name{PC1calling}
\alias{PC1calling}
\title{Compartments calling (A or B) from PC1 scores}
\usage{
PC1calling(bedgraph)
}
\arguments{
\item{bedgraph}{\code{GRanges} or \code{data.frame} object or bedgraph path (data frame without header and 4 columns tab separated).}
}
\value{
\code{GRanges}
}
\description{
From the scores of each bin (i.e first principal component scores), \code{PC1calling} output a \code{GRanges} with compartment A for score >= 0 and compartment B for score < 0
}
\details{
From bedgraph as input (\code{GRanges} object or path) output a \code{GRanges} with the domains A or B.
\code{NA} score are considered as part of a compartment A if they are between two compartments A (the other way around for B compartments).
Otherwise (if \code{NA} are between compartment A and B) they are not called, thus leaving a gap between two compartments.
}
\examples{
PC1calling(PC1_1_50kb.bedgraph)

}
