% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.geneTopology.R
\name{plot.geneTopology}
\alias{plot.geneTopology}
\title{plot method for "geneTopology" class}
\usage{
\method{plot}{geneTopology}(geneTopology_obj, TAD_id = NULL)
}
\arguments{
\item{geneTopology_obj}{A geneTopology object (ie output of genTADtopo function).}

\item{TAD_id}{TAD id as character}
}
\value{
Return two ggplot graph
}
\description{
plot method for "geneTopology" class
}
