% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rna_seq_chr19_10.1to10.6mb.gr}
\alias{rna_seq_chr19_10.1to10.6mb.gr}
\title{RNA seq coverage for Human chr 19 download on Ensembl
"http://ftp.ensembl.org/pub/release-104/bamcov/homo_sapiens/genebuild/GRCh38.illumina.colon.1.bam.bw.}
\format{
\subsection{\code{rna_seq_chr19_10.1to10.6mb.gr}}{

Grange with RNA coverage.
}
}
\usage{
rna_seq_chr19_10.1to10.6mb.gr
}
\description{
RNA seq coverage for Human chr 19 download on Ensembl
"http://ftp.ensembl.org/pub/release-104/bamcov/homo_sapiens/genebuild/GRCh38.illumina.colon.1.bam.bw.
}
\keyword{datasets}
