% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rna_seq_chr25_13to16mb.bedgraph}
\alias{rna_seq_chr25_13to16mb.bedgraph}
\title{RNA seq datas for chr 25 download on Ensembl
"http://ftp.ensembl.org/pub/release-104/bamcov/bos_taurus/genebuild/ARS-UCD1.2.ENA.heart.1.bam.bw).}
\format{
\subsection{\code{rna_seq_chr25_13to16mb.bedgraph}}{

Grange with mean coverage.
}
}
\usage{
rna_seq_chr25_13to16mb.bedgraph
}
\description{
RNA seq datas for chr 25 download on Ensembl
"http://ftp.ensembl.org/pub/release-104/bamcov/bos_taurus/genebuild/ARS-UCD1.2.ENA.heart.1.bam.bw).
}
\keyword{datasets}
