% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tad_2_10kb.bed}
\alias{tad_2_10kb.bed}
\title{Topological Associated Domain of individual 2 and chromosome 25/26.}
\format{
\subsection{\code{tad_2_10kb.bed}}{

A data frame containing TADs (domains) predicted by HiCExplorer on 10kb matrix.
For individual 2, chromosome 25 and 26.
\describe{
\item{chr}{chromosome name}
\item{start}{TAD start}
\item{end}{TAD end}
}
}
}
\usage{
tad_2_10kb.bed
}
\description{
Topological Associated Domain of individual 2 and chromosome 25/26.
}
\keyword{datasets}
