% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tad_HCT116_5kb.bed}
\alias{tad_HCT116_5kb.bed}
\title{TADs of Human HCT116 cell.}
\format{
\subsection{\code{tad_HCT116_5kb.bed}}{

A data frame containing TAD bounaries.
For autosomes from Human HCT116 cells with a bin width of 50kb.
Downloaded from the 4DN portal with accession number 4DNESNSTBMBY.
\describe{
\item{chr}{chromosome name}
\item{start}{boundary start}
\item{end}{boundary end}
}
}
}
\usage{
tad_HCT116_5kb.bed
}
\description{
TADs of Human HCT116 cell.
}
\keyword{datasets}
