% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TADarea.R
\name{TADarea}
\alias{TADarea}
\title{Annotations analysis around TAD boundaries}
\usage{
TADarea(tad.gr, annot.gr, window.size = 50000, tad.boundary = "start")
}
\arguments{
\item{tad.gr}{GRange file with TADs.}

\item{annot.gr}{GRange file with genomic annotations.}

\item{window.size}{Window (+/- value) in bp surrounding TAD boundary.}

\item{tad.boundary}{TAD border to be analyzed, "start", "end", and "center" denoting what to use as an anchor. Default is the start of each TADs.}
}
\value{
GRange file with all annotations surronding all TAD boundary (starts or stops...). TAD boundaries are at the window.size position.
}
\description{
For each TAD boundary (start or end), this function return all annotations that are in the desired window of the boundary.
The main use of this function is to be combined with areaHist or areaCov functions.
It is possible to use the start or the end of the TAD to analyze the distribution of annotations on both sides (note that it can be possible to use the center of the TAD).
This function can be time consuming so it is recommended to test it in a subset of the tad.gr.
}
\examples{
# tad.gr with 2 TADs
tad.gr <- dataframes2grange(
  data.frame(chr = rep(1, 2), start = c(2e6, 4e6), end = c(3e6, 5e6)),
  data.frame(chr = "1", size = 6e6)
)


# annot.gr with 3 genes
annot.gr <- dataframes2grange(
  data.frame(
    chr = rep(1, 3),
    start = c(1.5e6, 3.5e6, 5.5e6),
    end = c(1.6e6, 3.6e6, 5.6e6),
    type = c("a", "b", "c")
  ),
  data.frame(chr = "1", size = 6e6),
  metadata.mcols = 4
)

TADarea(tad.gr = tad.gr, annot.gr = annot.gr, window.size = 2e6)
}
