% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TADhist.R
\name{TADhist}
\alias{TADhist}
\title{Histogram of annotation distribution within TADs}
\usage{
TADhist(
  tad.gr,
  annot.gr,
  annot.border = "start",
  ifoverlap = "remove",
  annot.strand = FALSE,
  bin.width = 5,
  output = "plot"
)
}
\arguments{
\item{tad.gr}{GRange file with TADs.}

\item{annot.gr}{GRange file with genomic annotations.}

\item{annot.border}{Type of feature to analyzed. Start, end or center of each annotations from annot.gr files.}

\item{ifoverlap}{In case of annotation overlap a TAD boundary, few options are available:
(1)"remove" to remove all annot.gr that overlaps a TAD boundary,
(2)"real" to take the position of the annot.border in which it is located,
(3)"best" to take the position of the annot.border (that overlaps a TAD boundary) based on the TAD in which this annot.gr has the best overlay.}

\item{annot.strand}{Default is FALSE to plot the distribution as histogram. If TRUE, distributions are separated according to their strands and are displayed with lines.}

\item{bin.width}{Size of the bin in percent to count the number of annotations features, default is 5. if ifoverlap = "best": the real bin distances (before and after TAD boundaries) is equal to bin.width multiply by 1kb (therefore default bin size is 5kb).}

\item{output}{Default is "plot" to return a ggplot. Use "data" to return the datas used to produce the plot.}
}
\value{
Return a ggplot graph
}
\description{
Graph of the distribution of the genomic annotations within the TADs (relative distance)
Measures the relative distances of annotation border (start, end or the center) according to TAD size and plot the distribution as an histogram.
}
\details{
As an exemple, this function take all TAD domains and count the relative position of all annotation features (start, end or even the center).
It is possible that some annotations overlap a TAD boundary, in this cases few option are possible (see ifoverlap parameter and the exemple for better understanding):
(1)remove those annotations,
(2)uses the TAD in which the features (start, stop or center of the annotation) is located,
(3)uses the TAD in which the annotation has the largest overlap.
}
\examples{
#Create 20 genes
annot.gr <- dataframes2grange(
  data.frame(chr = 1, start = seq(190.5e3, 209.5e3, 1e3), end = 290e3, strand = "+"),
  data.frame(chr = "1", size = 400e3),
  strand.col = 4
)
#Create 2 TADs
tad.gr <- dataframes2grange(
  data.frame(chr = 1, start = c(100e3, 200e3), end = c(200e3, 300e3)),
  data.frame(chr = "1", size = 400e3)
)

TADplot(tad.gr = tad.gr, annot.gr = annot.gr, start = 150e3, stop = 300e3, chr = 1)

#see distribution of gene starts according to the "ifoverlap" parameters:
TADhist(tad.gr = tad.gr, annot.gr = annot.gr, ifoverlap = "real")
TADhist(tad.gr = tad.gr, annot.gr = annot.gr, ifoverlap = "remove")
TADhist(tad.gr = tad.gr, annot.gr = annot.gr, ifoverlap = "best")

}
