% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TADplot.R
\name{TADplot}
\alias{TADplot}
\title{Plot TADs for one individual}
\usage{
TADplot(
  tad.gr,
  chr,
  start,
  stop,
  tad.id = FALSE,
  bigwig.path = NULL,
  bigwig.binsize = 1000,
  bigwig.xaxis = "mean",
  bigwig.chr = NULL,
  bigwig.yaxis = NULL,
  annot.gr = NULL,
  annot.col = NULL,
  bedgraph.path = NULL
)
}
\arguments{
\item{tad.gr}{GRange object with TADs. This file must have chromosomes lengths (see dataframes2grange function).}

\item{chr}{Chromosome name to plot.}

\item{start}{Start of the area of interest.}

\item{stop}{End of the area of interest.}

\item{tad.id}{Default is FASLE to write the size of each TAD instead of their names.}

\item{bigwig.path}{Path for the bigwig file plotted as histogram. Default = NULL (ie no track is plotted).}

\item{bigwig.binsize}{Bin sizes for the histogram of the bigwig track. Default = 1e3.}

\item{bigwig.xaxis}{Function used to transforming the x-axis of bigwig values among each bigwig.binsize. Defaults = "median".
Alternatively, other predefined functions can be supplied as character (mean, median, sum, min, max or extreme).}

\item{bigwig.chr}{Chromosome name used for the bigwig file(s). Default = NULL to used the same name as chr.}

\item{bigwig.yaxis}{Function used to transforming the y-axis of bigwig values. Default = NULL. Use "log2" to use the function log2(x + 1) to transform the y-axis or provide any other function.}

\item{annot.gr}{GRange file with genomic annotations. Default = NULL (ie no track is plotted).}

\item{annot.col}{Column number of the metadata from annot.gr files used to group the annotation tracks. Default = NULL.}

\item{bedgraph.path}{Path for the bedgraph file plotted as line. Default = NULL (ie no track is plotted).}
}
\value{
Plot with TADs and other tracks as a list of GenomeGraph tracks (see Gviz::plotTracks for details).
}
\description{
TADplot use Gviz package to plot at least 2 tracks. The first one includes all the TADs from the specified chromosome. And the second one is a zoom in the defined area.
Other tracks can be added like gene annotation or read depth data.
}
\details{
This create a plot with at least 2 tracks from TADs annotation (tad.gr).

On the first track the alternating TADs are represented in grey and black while the interTAD regions are represented in white.
The second track is a zoom in the area of interest. This area is extend to the border of the TADs on both sides.

Another track from a bigwig file can be added like read depth sequencing (RNAseq...) as an histogram.
The bin size of the histogram is 1Kb by default, the read depth is smoothed using the median value of each bin.
If the chromosome name is different (between the bigwig and the chr parameter), it can be fix using the bigwig.chr parameter ("1" versus "chr1").

Another track with any annotations can be added.
This track can be group using factors in a specified column of the annot.gr files (metadata) otherwise the names of each annotation is used.
}
\examples{
# Lets create a GRange object with 3 TADs and 1 metadata column:
tad.gr <- dataframes2grange(
  data.frame(
    chr = rep(1, 3),
    start = c(2e6, 3e6, 4e6),
    end = c(3e6, 4e6, 5e6),
    type = c("a", "b", "a")
  ),
  data.frame(chr = "1", size = 6e6),
  metadata.mcols = 4
)
tad.gr

# plot those TADs
# use those TADs as annotation tracks (with the metadata column to group annotations):
TADplot(
  tad.gr = tad.gr,
  chr = 1, start = 2e6, stop = 5e6,
  annot.gr = tad.gr, annot.col = 1
)
}
