% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TADshuffling.R
\name{TADshuffling}
\alias{TADshuffling}
\title{TAD shuffling}
\usage{
TADshuffling(tad.gr)
}
\arguments{
\item{tad.gr}{GRange object with TADs. This file must have chromosomes lengths (see dataframes2grange function).}
}
\value{
GRange file
}
\description{
From a set of TADs, this function return un random set of TADs.
}
\details{
This function shuffle the TADs order for each chromosomes (i.e. keep the TADs sizes). The TADs input should not have gaps (region between two TADs) excepted for the extremities of chromosomes.
}
