% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areaCov.R
\name{areaCov}
\alias{areaCov}
\title{Density plot of annotations around boundaries}
\usage{
areaCov(data.gr, annot.strand = TRUE, bin.width = NULL, norm = FALSE)
}
\arguments{
\item{data.gr}{Output of \code{boundArea()}.}

\item{annot.strand}{Logical. If \code{TRUE} (default): separate coverage according to their strands.}

\item{bin.width}{Size of the sliding window in base pair to calculate the mean coverage. Default is \code{NULL} to to use a size in 10 time smaller than \code{window.size} parameter of \code{boundArea()}.}

\item{norm}{Logical. Normalized density using relative content (zscore of density). Default is \code{FALSE}.}
}
\value{
\code{ggplot}
}
\description{
This function take the output of \code{boundArea()} to produce a graph of the density of the genomic annotations around TAD boundaries (real distance).
}
\details{
If the density is significantly different between strands, it can be useful to use relative density (i.e normalized using the zscore with norm = TRUE parameter).
}
\examples{
# create GRange with 1 TAD
tad.gr <- dataframes2grange(
  data.frame(chr = 1, start = c(10e3), end = c(50e3)),
  data.frame(chr = "1", size = 60e3)
)

# create annotations of 12 genes every 2kb with width = 10kb on both strand:
annot.gr <- dataframes2grange(
  data.frame(
    chr = rep("1", 12),
    start = seq(20e3, 42e3, 2e3),
    end = seq(20e3, 42e3, 2e3) + 10e3,
    strand = c(rep("+", 6), rep("-", 6))
  ),
  data.frame(chr = "1", size = 60e3),
  strand.col = 4
)
# Vizualisation
TADplot(tad.gr = tad.gr, start = 5e3, stop = 55e3, chr = 1, annot.gr = annot.gr)

# Distribution analysis surrounding the start of the TAD
data.gr <- boundArea(
  domain.gr = tad.gr,
  annot.gr = annot.gr,
  window.size = 60e3,
  domain.boundary = "start"
)

# plot coverage
areaCov(data.gr = data.gr, annot.strand = FALSE)
areaCov(data.gr = data.gr, annot.strand = TRUE)

}
