% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary2domain.R
\name{boundary2domain}
\alias{boundary2domain}
\title{Convert boundary to domain}
\usage{
boundary2domain(boundaries)
}
\arguments{
\item{boundaries}{data frame (chromosome, start and end) or GRanges with the boundaries}
}
\value{
GRanges object with boundaries
}
\description{
From data with each boundaries stored by line (chromosome, start and end), return GRanges with domains.
}
\details{
Start or end position of domains are the middle of a boundary.
}
