% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw2grange.R
\name{bw2grange}
\alias{bw2grange}
\title{Convert bigwig file to \code{GRanges} object}
\usage{
bw2grange(bigwig.path, bin.width = NULL)
}
\arguments{
\item{bigwig.path}{path of bigwig file.}

\item{bin.width}{Bin width. Default is NULL to return the bin size as it is in the bigwig file.}
}
\value{
\code{GRanges}
}
\description{
From the full path of a bigwig files, return a \code{GRanges} object. The bin width of the \code{GRanges} can be specify with \code{bin.width} parameter.
}
