% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chromsize}
\alias{chromsize}
\title{Size of cow autosomes.}
\format{
\subsection{\code{chromsize}}{

A data frame containing the size in base pair of bovine chromosomes
\describe{
\item{chr}{chromosome name}
\item{size}{sine in base pair}
}
}
}
\usage{
chromsize
}
\description{
Size of cow autosomes.
}
\keyword{datasets}
