% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compOrientation.R
\name{compOrientation}
\alias{compOrientation}
\title{Orientation of PC1 score}
\usage{
compOrientation(bedgraph.gr, annot.gr, expression.data.frame)
}
\arguments{
\item{bedgraph.gr}{\code{GRanges} file with the score to be used for compartment calling (i.e PC1 values).}

\item{annot.gr}{\code{GRanges} file with gene annotations.}

\item{expression.data.frame}{\code{dataframe} with 3 columns:
\itemize{
\item 1: gene IDs: must be identical IDs than \code{names(annot.gr)},
\item 2: common name for genes: could be the gene ID,
\item 3: expression of each gene: such as raw count or \code{log(raw_count + 1)}.
}}
}
\value{
S3 class object with 3 \code{dataframes}.
}
\description{
Considering that the A compartments have a stronger expression than the B compartments.
This function allows the score (i.e PC1 score) to be oriented (positif or negatif).
Then the compartments A (actif) and B (inactif) can be called accurately.
}
\details{
This function is divided into 3 steps:
\itemize{
\item Compartments calling from the score (see PC1calling function).
\item Measures median gene expression in compartments A and B per chromosome.
\item Reverse the score if median expression of A is smaller than B compartments.
}

Return a list with 3 \code{dataframe}.
\itemize{
\item bedgraph as \code{GRanges} object with the score oriented.
\item \code{dataframe} with expression of each gene (before orientation).
\item \code{dataframe} with median expression of each compartment (before orientation).
}
}
\examples{
# see vignette("Turorial_TADkit_R_package") or on github (https://github.com/Nico-FR/TADkit)

}
