% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coolFetch.R
\name{coolFetch}
\alias{coolFetch}
\title{Fetch a chromosome of a cool or mcool file}
\usage{
coolFetch(cool.path, chr, bin.width = NA, balance = FALSE)
}
\arguments{
\item{cool.path}{The full path of the cool or mcool file.}

\item{chr}{The selected chromosome.}

\item{bin.width}{Bin width (i.e resolution) of the mcool matrix in bp. It must not be used for cool file.}

\item{balance}{Logical. Weather or not to use balanced counts instead of raw counts. Default = \code{FALSE}.}
}
\value{
A \code{dgCMatrix} object: upper triangular and sparse Matrix
}
\description{
From cool or mcool files, \code{coolFetch} import the interaction matrix for one chromosome.
If balance = \code{TRUE} returns the normalized counts
}
\details{
Using the Indexes provided by the cooler data model :
https://cooler.readthedocs.io/en/latest/schema.html,
extract the intra-chromosomal counts for a given chromosome
}
