% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coolFetch.R
\name{coolFetch}
\alias{coolFetch}
\title{Fetch a chromosome of a mcool file}
\usage{
coolFetch(path, chr, bin.width = NA, balance = FALSE)
}
\arguments{
\item{path}{the full path of the cool or mcool file}

\item{chr}{the selected chromosome}

\item{bin.width}{bin width (i.e resolution) of the mcool matrix in bp}

\item{balance}{logical. Weather or not to use balanced counts instead of raw counts. Default = FALSE}
}
\value{
a sparse matrix: an object sparseMatrix
}
\description{
Using the Indexes provided byt the cooler data model :
https://cooler.readthedocs.io/en/latest/schema.html,
extract the intra-chromosomal counts for a given chromosome
If balance = TRUE returns the normalized counts
}
