% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coolFetch.R
\name{coolFetch}
\alias{coolFetch}
\title{Fetch a chromosome of a cool or mcool file}
\usage{
coolFetch(cool.path, chr, bin.width = NA, balance = FALSE)
}
\arguments{
\item{cool.path}{The full path of the cool or mcool file.}

\item{chr}{The selected chromosome.}

\item{bin.width}{Bin width in base pair (i.e resolution) for mcool file. Default is \code{NA} for cool file.}

\item{balance}{Logical. Weather or not to use balanced counts instead of raw counts. Default = \code{FALSE}.}
}
\value{
A \code{dgCMatrix} object: upper triangular and sparse Matrix
}
\description{
From cool or mcool files, \code{coolFetch} import the interaction matrix for one chromosome.
If balance = \code{TRUE}, \code{coolFetch} returns the normalized counts
}
\details{
The cool file format is an efficient storage format for high resolution genomic interaction matrices, developed and maintained by the Mirny lab (https://github.com/open2c/cooler)
\code{coolFetch} use the indexes provided by the cooler data model to extract the intra-chromosomal counts for a given chromosome.
As cool files store genomic interactions for only one resolution, \code{bin.width} must be set to \code{NA}.
While, as mcool files store genomic interactions for multiples resolutions, the chosen resolution must be set with the \code{bin.width} parameter.
}
