% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframes2grange.R
\name{dataframes2grange}
\alias{dataframes2grange}
\title{Create GRanges with metadatas}
\usage{
dataframes2grange(
  annotation.table,
  chromsize,
  chr.col = 1,
  start.col = 2,
  end.col = 3,
  strand.col = NULL,
  name.col = NULL,
  metadata.mcols = NULL
)
}
\arguments{
\item{annotation.table}{\code{dataframe} with genomic annotations (at least 3 columns: chromosome, start and stop).}

\item{chromsize}{\code{dataframe} with chromosome names and sizes. Chromosomes are also filtered according to this list.}

\item{chr.col}{Column number for chromosome names, default is 1.}

\item{start.col}{Column number for start, default is 2.}

\item{end.col}{Column number for end, default is 3.}

\item{strand.col}{Column number for strands, default is \code{NULL} to set strands as "*" (i.e both strands or unknown stand) otherwise the column must contain "+" or "-" characters.}

\item{name.col}{Column number for annotation names, default is \code{NULL} to create names with chr.names and start.}

\item{metadata.mcols}{Column number(s) for metadata, default is \code{NULL} (i.e no metadata).}
}
\value{
GRange file that contains chromosome size.
}
\description{
This function create a GRange object from the first \code{dataframe} (annotation.table) and add the size of the chromosomes on the second \code{dataframe}.
The chromosomes are filtered according to the list of chromosomes contained in the second \code{dataframe} (chromsize).
}
\examples{
bed.df = data.frame(chr = 1, start = 198e3, end = 290e3, strand = "+", names = "geneID")
chromsize = data.frame(chr = "1", size = 400e3)
dataframes2grange(bed.df, chromsize, strand.col = 4,name.col = 5)
}
