% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneTADtopo.R
\name{geneTADtopo}
\alias{geneTADtopo}
\title{Genes topology according to TAD}
\usage{
geneTADtopo(tad.gr, gene.gr, ifoverlap = "best", expression.data.frame = NULL)
}
\arguments{
\item{tad.gr}{\code{GRanges} object with domains. The chromosomes lengths must be stored as metadata (see \code{dataframes2grange()}).}

\item{gene.gr}{\code{GRanges} with genomic annotations.}

\item{ifoverlap}{In case of annotation overlap a TAD boundary, few options are available to measure the \code{annot.border} positions:
\itemize{
\item \code{"remove"} to remove all \code{annot.gr} that overlaps a TAD boundary,
\item \code{"real"} to take the position according to the TAD where the \code{annot.border} is located,
\item \code{"best"} to take the position according to the TAD where the \code{annot.gr} has the best overlay.
}}

\item{expression.data.frame}{\code{dataframe} with expression data (raw counts...).
The first two columns should be used to identify the genes. The 1st one must use the same gene id than \code{names(gene.gr)} (i.e unique ID). The 2nd can be used with usual gene names.
Others columns give expression count, one column per sample/experiment.}
}
\value{
Return a S3 class object with 3 data frames.
}
\description{
For each domain (e.g TAD) the function return a data frame with the numbers, names and strands of genes. It also return some statistics which can be plotted using plot(output).
}
\details{
Return a S3 object with 3 dataframes. The second and the third dataframe can be plotted using the plot function.
}
\examples{
# output <- geneTADtopo(tad.gr, gene.gr)
# plot(output)

}
