% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matObsExp.R
\name{matObsExp}
\alias{matObsExp}
\title{observed / expected matrix}
\usage{
matObsExp(matrix)
}
\arguments{
\item{matrix}{a dgCMatrix or matrix object}
}
\value{
a dgCMatrix object: upper triangular and sparse Matrix
}
\description{
For each bin of the matrix (interaction count observed) this function return a ratio: observed / expected.
The input matrix must be a dgCMatrix or matrix object (i.e array with as many rows and columns than the number of bin). Only the upper part of the matrix is used.
The output can be plot with MATplot function with log2=T and scale.colors = "ObsExp" parameters.
}
\details{
The expected number of interaction corresponds to the average of the interaction counts according to the bin distances (NB: bins with zero counts are not considered).
}
