% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matObsExp.R
\name{matObsExp}
\alias{matObsExp}
\title{observed / expected matrix}
\usage{
matObsExp(matrix)
}
\arguments{
\item{matrix}{\code{dgCMatrix} or \code{matrix} object.}
}
\value{
\code{dgCMatrix} object: upper triangular and sparse Matrix
}
\description{
For each bin of the matrix (interaction count observed) \code{matObsExp()} return the ratio: observed / expected.
The input matrix must be a \code{dgCMatrix} or \code{matrix} object (i.e matrix with as many rows and columns than the number of bin).
The output can be plot with \code{MATplot(output, log2 = TRUE, scale.colors = "ObsExp")}.
}
\details{
The expected number of interaction corresponds to the average of the interaction counts according to the bin distances.
Bins with zero counts are not considered.
Only the upper part of the matrix is used.
}
