% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matObsExp.R
\name{matObsExp}
\alias{matObsExp}
\title{observed / expected matrix}
\usage{
matObsExp(matrix, output = "OE")
}
\arguments{
\item{matrix}{\code{Matrix} or \code{matrix} object.}

\item{output}{default is "OE" to return observed / expected matrix. Use "E" or "Exp" to return expected matrix.}
}
\value{
\code{dgCMatrix} object: upper triangular and sparse Matrix
}
\description{
For each bin of the matrix (interaction count observed) \code{matObsExp()} return the ratio: observed / expected.
The input matrix must be a \code{Matrix} or \code{matrix} object (i.e matrix with as many rows and columns than the number of bin).
The output can be plot with \code{MATplot(output, log2 = TRUE, scale.colors = "ObsExp")}.
}
\details{
The expected number of interaction corresponds to the average interaction counts according to bin distances.
Note that the expected number of interaction is only estimated from the chromosome supplied. Other genome wide approaches may be considered.
}
\examples{
mat_obsexp = matObsExp(mat_HCT116_chr19_50kb)
MATplot(matrix = mat_obsexp,
    start = 5e6, stop = 15e6,
    bin.width = 50e3,
    log2 = TRUE, scale.colors = "OE")

}
