% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MATplot.R
\name{MATplot}
\alias{MATplot}
\title{Plot TADs for one individual}
\usage{
MATplot(
  matrix,
  start,
  stop,
  bin.width,
  matrix.colname = T,
  matrix.rowname = T,
  matrix.sep = "\\t",
  matrix.diag = T,
  log2 = T,
  scale.colors = "H"
)
}
\arguments{
\item{matrix}{matrix object (data frame or matrix) or matrix path at matrix format (dense) for only one chromosome. The path can be gzip (ending by .gz)}

\item{start}{start in bp of the area of interest.}

\item{stop}{end in bp of the area of interest.}

\item{bin.width}{bin width of the matrix in bp.}

\item{matrix.colname}{logical. Does your matrix file (ie path) have column names (ie header)? Default = TRUE}

\item{matrix.rowname}{logical. Does your matrix file (ie path) have row names? Default = TRUE}

\item{matrix.sep}{the field separator character. Values on each line of the matrix file are separated by this character. Default = '\\t' (ie tabulation).}

\item{matrix.diag}{logical. Weather or not to plot diagonal of the matrix. Default = TRUE}

\item{log2}{logical. Weather or not to plot the log2 of the matrix values.}

\item{scale.colors}{A character string indicating the color map option to use. Eight options are available (see viridis package:
"magma" (or "A")
"inferno" (or "B")
"plasma" (or "C")
"viridis" (or "D")
"cividis" (or "E")
"rocket" (or "F")
"mako" (or "G")
"turbo" (or "H")}
}
\value{
ggplot of the matrix.
}
\description{
TADplot use Gviz package to plot at least 2 tracks. The first one includes all the TADs from the specified chromosome. And the second one is a zoom in the defined area.
Other tracks can be added like gene annotation or read depth data.
}
\details{
This create a plot with at least 2 tracks from TADs annotation (tad.gr).

On the first track the alternating TADs are represented in grey and black while the interTAD regions are represented in white.
The second track is a zoom in the area of interest. This area is extend to the border of the TADs on both sides.

Another track from a bigwig file can be added like read depth sequencing (RNAseq...) as an histogram.
The bin size of the histogram is 1Kb by default, the read depth is smoothed using the median value of each bin.
If the chromosome name is different (between the bigwig and the chr parameter), it can be fix using the bigwig.chr parameter ("1" versus "chr1").

Another track with any annotations can be added.
This track can be group using factors in a specified column of the annot.gr files (metadata) otherwise the names of each annotation is used.
}
