% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MATplot.R
\name{MATplot}
\alias{MATplot}
\title{Plot matrix with annotations}
\usage{
MATplot(
  matrix,
  start,
  stop,
  bin.width,
  log2 = T,
  scale.colors = "H",
  matrix.diag = T,
  tad.upper.tri = NULL,
  tad.lower.tri = NULL,
  loop.bedpe = NULL,
  tad.chr = NULL,
  annotations.color = "red",
  tad.upper.line = NULL,
  tad.lower.line = NULL,
  tad.line.col = NULL,
  line.colors = c("red", "blue")
)
}
\arguments{
\item{matrix}{matrix object (of class "dgCMatrix" or "matrix") for only one chromosome.}

\item{start, stop}{region of interest in base pair.}

\item{bin.width}{bin width of the matrix in base pair.}

\item{log2}{logical. Use the log2 of the matrix values. Default is TRUE}

\item{scale.colors}{A character string indicating the color map option to use. Eight colors palettes are available from viridis package. Another palette "OE" is made for data centered on 0 (i.e log2 of observed/expected matrix). Default is "H":
\itemize{
\item "ObsExp" (or "OE"),
\item "magma" (or "A"),
\item "inferno" (or "B"),
\item "plasma" (or "C"),
\item "viridis" (or "D"),
\item "cividis" (or "E"),
\item "rocket" (or "F"),
\item "mako" (or "G"),
\item "turbo" (or "H").
}}

\item{matrix.diag}{logical. Weather or not to plot diagonal values of the matrix. Default = TRUE}

\item{tad.upper.tri, tad.lower.tri}{bed files path, data frame or grange object with the TAD to plot as triangle in the upper or lower part of the matrix. Default is NULL.}

\item{loop.bedpe}{bedpe files path or data frame to plot on both parts of the matrix. Six columns table (chr1, start1, end1, chr2, start2, end2) that gives loops between 2 areas. Default is NULL}

\item{tad.chr}{chromosome name to filter bed and bedpe files. Default is NULL}

\item{annotations.color}{color for loop and tri annotations. Default is "red".}

\item{tad.upper.line, tad.lower.line}{bed files path, data frame or grange object with the TAD to plot as line in the upper or lower parts of the matrix. Default is NULL.}

\item{tad.line.col}{column number of the tad.line files that contain factors used to color tad.upper.line and tad.lower.line. Default is NULL.}

\item{line.colors}{colors for upper and lower lines. Default is red and blue.}
}
\value{
ggplot
}
\description{
MATplot allow to plot matrix with 3 types of annotations:
-domains (e.g. TADs or compartments) are plot as triangles or lines on the upper or/and lower part of the matrix.
-interactions between domains/bins (loop) are plot as squares on the upper and lower part of the matrix.
}
\details{
The matrix input must be a dgCMatrix or a matrix object for only one chromosome (see coolFetch function to read cool files).
All domains (TADs or compartments) are bed files (3 columns: chr, start and end) and can be R object (dataframe or GRange) or the path of the files.
For tad.lines, another column can be used to specify different classes of domains (e.g compartment A or B). To use those domain classes, specify the column number (of the tad.upper.line and tad.lower.line inputs) with tad.line.col parameter and a custom set of colors with line.colors parameter.
Loop are stored in bedpe files (6 columns: chr1, start1, end, chr2, start2 and end2) and can be a dataframe or the path of the file.
Chromosome domains and loops can be filter using tad.chr parameter.
}
