% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{matrix_1_chr26_50kb}
\alias{matrix_1_chr26_50kb}
\title{Matrix of bovine 1 and chromosome 26.}
\format{
\subsection{\code{matrix_1_chr26_50kb}}{

A dgCMatrix object with 1040 rows and columns containing the raw counts (number of interactions).
For bovine 1, chromosome 26 with a bin width of 50e3 bases pairs.
Only the upper matrix and non zero count are stored.
}
}
\usage{
matrix_1_chr26_50kb
}
\description{
Matrix of bovine 1 and chromosome 26.
}
\keyword{datasets}
