% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca2matrix.R
\name{orca2matrix}
\alias{orca2matrix}
\title{Import Orca matrix into full chromosome matrix}
\usage{
orca2matrix(
  df_prediction.path,
  sep = " ",
  mpos,
  scale,
  chromsize,
  output = "OE",
  df_normmats.path = NULL
)
}
\arguments{
\item{df_prediction.path}{The full path of the Orca matrix. Data frame without header which contains orca predictions, i.e. log(Obseved/Expected).}

\item{mpos}{The coordinate to zoom into for multiscale prediction.}

\item{scale}{Size of the scale (i.e. window) of the predictions in base pair (i.e. 32Mb, 16Mb, 8Mb, 4Mb, 2Mb and 1Mb).}

\item{chromsize}{Chromosome size in base pair of the related chromosome.}

\item{output}{Character: Default is "OE" to return observed / expected counts. "Obs" to return observed counts.}

\item{df_normmats.path}{Optional. Background distance-based expected balanced contact matrices. Only needed to return observed counts.}
}
\value{
A \code{dgCMatrix} object: upper triangular and sparse Matrix
}
\description{
This function create an empty matrix of the full chromosome which is then filled by Orca matrix (i.e. 250x250 bins).
}
\details{
From the matrices produced by orca_predict.genomepredict function.
}
