% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PC1calling.R
\name{PC1calling}
\alias{PC1calling}
\title{Compartments calling (A or B)}
\usage{
PC1calling(bedgraph)
}
\arguments{
\item{bedgraph}{GRange file or bedgraph path (4 columns: chr, star, end, pc1 score) for PC1 values.}
}
\value{
grange
}
\description{
From a score for each bin (i.e principal component), the function output a Grange file with compartment A for score >= 0 and compartment B for score < 0
}
\details{
From bedgraph as input (GRange object or path) output a GRange with the compartments.
NA score are considered as part of a compartment A if they are between two compartments A (the other way around for B compartments).
Otherwise (if NA are between compartment A and B) they are not called, thus leaving a gap between compartments.
}
\examples{
# output <- geneTADtopo(tad.gr, gene.gr)
# plot(output)

}
