% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tad_1_10kb.bed}
\alias{tad_1_10kb.bed}
\title{TAD of bovine 1.}
\format{
\subsection{\code{tad_1_10kb.bed}}{

A data frame containing TADs (domains) predicted by HiCExplorer on 10kb matrix.
\describe{
\item{chr}{chromosome name}
\item{start}{TAD start}
\item{end}{TAD end}
}
}
}
\usage{
tad_1_10kb.bed
}
\description{
TAD of bovine 1.
}
\keyword{datasets}
