% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TADarea.R
\name{TADarea}
\alias{TADarea}
\title{Annotations analysis around TAD boundaries}
\usage{
TADarea(tad.gr, annot.gr, window.size = 50000, tad.boundary = "start")
}
\arguments{
\item{tad.gr}{\code{GRanges} object with domains. The chromosomes lengths must be stored as metadata (see \code{dataframes2grange()}).}

\item{annot.gr}{\code{GRanges} object with genomic annotations..}

\item{window.size}{Window in bp surrounding each side of the TAD boundaries.}

\item{tad.boundary}{TAD border to be analyzed; \code{"start"}, \code{"end"} or \code{"center"} denoting what to use as an anchor. Default is the \code{"start"} of each TAD.}
}
\value{
\code{GRanges} object with all annotations surrounding all TAD boundaries. TAD boundaries are at the window.size position.
}
\description{
For each TAD boundary (start or end), \code{TADarea} return all annotations that are in the desired window of the boundary (i.e boundary +/- \code{window.size}).
The main use of this function is to be combined with \code{areaHist()} or \code{areaCov()} functions.
It is possible to use the \code{"start"}, \code{"end"} or \code{"center"} of the TAD to analyze the distribution of annotations.
}
\examples{
# tad.gr with 2 TADs
tad.gr <- dataframes2grange(
  data.frame(chr = rep(1, 2), start = c(2e6, 4e6), end = c(3e6, 5e6)),
  data.frame(chr = "1", size = 6e6)
)


# annot.gr with 3 genes
annot.gr <- dataframes2grange(
  data.frame(
    chr = rep(1, 3),
    start = c(1.5e6, 3.5e6, 5.5e6),
    end = c(1.6e6, 3.6e6, 5.6e6),
    type = c("a", "b", "c")
  ),
  data.frame(chr = "1", size = 6e6),
  metadata.mcols = 4
)

TADarea(tad.gr = tad.gr, annot.gr = annot.gr, window.size = 2e6)
}
