% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TADdiff.R
\name{TADdiff}
\alias{TADdiff}
\title{TAD boundaries differences between 2 or more samples.}
\usage{
TADdiff(
  boundaries.lst,
  score.lst,
  bin.width = NULL,
  window.size = NULL,
  restrict = 2e+06
)
}
\arguments{
\item{boundaries.lst}{List of TAD boundaries as \code{GRanges}. Each \code{GRanges} in the list must have names and seqlengths metadata. The width of each boundary must correspond to the resolution of the matrix used and must be the same between samples.}

\item{score.lst}{List of insulation score as a \code{GRanges}. Each \code{GRanges} in the list must have names and seqlengths metadata.}

\item{bin.width}{Default is \code{NULL} to estimate the bin.width from \code{boundaries.lst} files.}

\item{window.size}{Defaults is \code{NULL} to take a value 2 times higher than \code{bin.width}, otherwise set the \code{window.size} in base pairs.}

\item{restrict}{Default is \code{2e6} (i.e 2Mb) to remove boundary at the extremities of chromosomes.}
}
\value{
list of GRanges
}
\description{
\code{TADdiff} check if for each TAD boundary from one sample there is an other boundary in the vicinity on all other samples provided.
It also return the score for each boundary and the delta between each sample.
}
\details{
\code{TADdiff()} takes the score used to find TAD boundaries (e.g insulation score) and the TAD boundaries as a \code{GRanges} object (see \code{dataframes2grange()} function).
For each TAD boundary :
-checks if there is an other boundary in the other sample +/- \code{window.size} (e.g +/- 20kb if \code{bin.size = 10e3} and \code{window.size = NULL}),
-exports the score at the boundary and the delta score at he boundary compared to the other sample. Those delta scores can be used to rank and find the boundary with the highest differences.
Differences analysis (new boundary and delta score at boundary) are performed between each samples (2 samples or more).
Because the scores are often not accurate at the extremities of the chromosomes (beginning and the end of each chromosome), these regions can be ignored using \code{restrict} parameters.
}
