% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TADhist.R
\name{TADhist}
\alias{TADhist}
\title{Histogram of annotation distribution within domains}
\usage{
TADhist(
  tad.gr,
  annot.gr,
  annot.border = "start",
  ifoverlap = "remove",
  annot.strand = FALSE,
  bin.width = 5,
  output = "plot"
)
}
\arguments{
\item{tad.gr}{\code{GRanges} object with domains. The chromosomes lengths must be stored as metadata (see \code{dataframes2grange()}).}

\item{annot.gr}{\code{GRanges} object with genomic annotations.}

\item{annot.border}{Type of feature to analyzed. \code{"start"}, \code{"end"} or \code{"center"} of each annotations from \code{annot.gr} object.}

\item{ifoverlap}{In case of annotation overlap a TAD boundary, few options are available to measure the \code{annot.border} positions:
\itemize{
\item \code{"remove"} to remove all \code{annot.gr} that overlaps a TAD boundary,
\item \code{"real"} to take the position according to the TAD where the \code{annot.border} is located,
\item \code{"best"} to take the position according to the TAD where the \code{annot.gr} has the best overlay.
}}

\item{annot.strand}{Default is \code{FALSE} to plot the distribution as histogram. If \code{TRUE}, distributions are separated according to their strands and are displayed with lines.}

\item{bin.width}{Size of the bin in percent to count the number of annotations features, default is 5\%. if \code{ifoverlap = "best"}: the real bin distances (ploted before and after TAD boundaries) is equal to \code{bin.width*1e3}. Therefore default real bin size is 5kb.}

\item{output}{Default is \code{"plot"} to return a \code{ggplot}. Use \code{"data"} to return the datas used to produce the plot.}
}
\value{
\code{ggplot} object.
}
\description{
Plot of the distribution of the genomic annotations within the TADs (relative distance)
Measures the relative distances of annotation border (start, end or the center) according to TAD size and plot the distribution as an histogram.
}
\details{
As an example, \code{TADhist()} take all TAD domains and count the relative position of all annotation features (start, end or even the center).
It is possible that some annotations overlap a TAD boundary, in this cases few option are possible to get the distributions (see \code{ifoverlap} parameter and the example for better understanding):
* remove: those annotations,
* uses: the TAD in which the features (\code{"start"}, \code{"end"} or \verb{"center" of the annotation) is located, * uses: the TAD in which the annotation has the best overlap. Therefore, in some cases, the }"start"\verb{of an}annot.gr\verb{can be located before the TAD (i.e the TAD with the best overlap). In that case, the distance between the}"start"\code{of}annot.border` and TAD corresponds to the real distance (in base pair).
}
\examples{
#Create 20 genes
annot.gr <- dataframes2grange(
  data.frame(chr = 1, start = seq(190.5e3, 209.5e3, 1e3), end = 290e3, strand = "+"),
  data.frame(chr = "1", size = 400e3),
  strand.col = 4
)
#Create 2 TADs
tad.gr <- dataframes2grange(
  data.frame(chr = 1, start = c(100e3, 200e3), end = c(200e3, 300e3)),
  data.frame(chr = "1", size = 400e3)
)

TADplot(tad.gr = tad.gr, annot.gr = annot.gr, start = 150e3, stop = 300e3, chr = 1)

#see distribution of gene starts according to the "ifoverlap" parameters:
TADhist(tad.gr = tad.gr, annot.gr = annot.gr, ifoverlap = "real")
TADhist(tad.gr = tad.gr, annot.gr = annot.gr, ifoverlap = "remove")
TADhist(tad.gr = tad.gr, annot.gr = annot.gr, ifoverlap = "best")

}
