% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TADplot.R
\name{TADplot}
\alias{TADplot}
\title{Plot domains for one individual}
\usage{
TADplot(
  tad.gr,
  chr,
  start,
  stop,
  tad.id = FALSE,
  bigwig.path = NULL,
  bigwig.binwidth = 1000,
  bigwig.xaxis = "mean",
  bigwig.chr = NULL,
  bigwig.yaxis = NULL,
  annot.gr = NULL,
  annot.col = NULL,
  bedgraph.path = NULL
)
}
\arguments{
\item{tad.gr}{\code{GRanges} object with domains. The chromosomes lengths must be stored as metadata (see \code{dataframes2grange()}).}

\item{chr}{Chromosome name to plot.}

\item{start, stop}{Region of interest in base pair.}

\item{tad.id}{Logical. Default is \code{FALSE} to label the domain with their sizes \code{width(tad.gr)} instead of their names \code{names(tad.gr)}.}

\item{bigwig.path}{Path for the bigwig file plotted as histogram. Default = \code{NULL} (i.e the track is not plotted).}

\item{bigwig.binwidth}{Bin sizes for the histogram of the bigwig track. Default = \code{1e3}.}

\item{bigwig.xaxis}{Function used to transform the x-axis of the bigwig values among each bigwig.binwidth. Defaults = \code{"median"}.
Alternatively, other predefined functions can be supplied as character (\code{"mean"}, \code{"median"}, \code{"sum"}, \code{"min"}, \code{"max"} or \code{"extreme"}).}

\item{bigwig.chr}{Chromosome name used for the bigwig file. Default = \code{NULL} to used the same name as chr.}

\item{bigwig.yaxis}{Function used to transforming the y-axis of bigwig values. Default = \code{NULL}. Use \code{"log2"} to use the function \code{log2(x + 1)} to transform the y-axis or provide any other function.}

\item{annot.gr}{\code{GRanges} object with genomic annotations. Default = \code{NULL} (i.e the track is not plotted).}

\item{annot.col}{Column number of the metadata from \code{annot.gr} file used to group the annotation tracks. Default = \code{NULL}.}

\item{bedgraph.path}{Path for the bedgraph file plotted as line. Default = \code{NULL} (i.e the track is not plotted).}
}
\value{
Plot with domains and other tracks as a list of GenomeGraph tracks (see \code{Gviz::plotTracks} for details).
}
\description{
\code{TADplot()} use \code{Gviz} package to plot at least 2 tracks:
\itemize{
\item All the TADs from a specified chromosome,
\item A zoom in the defined area.
Other tracks can be added:
\item bed files with annotations (genes...),
\item bigwig files (read depth sequencing...),
\item bedgraph with bin values (insulation score...).
}
}
\details{
This create a plot with at least 2 tracks from TADs annotation (tad.gr).

On the first track the alternating TADs are represented in grey and black while the interTAD regions are represented in white.
The second track is a zoom in the area of interest. This area is extend to the border of the TADs on both sides.

Another track from a bigwig file can be added like read depth sequencing (RNAseq...) as an histogram.
The bin size of the histogram is 1Kb by default, the read depth is smoothed using the median value of each bin.
If the chromosome name is different (between the bigwig and the \code{chr} parameter), it can be fix using the \code{bigwig.chr} parameter (e.g \code{chr="1"} while \code{bigwig.chr="chr1"}).

Another track with any annotations can be added.
This track can be group using factors in a specified column of the \code{annot.gr} files (metadata) otherwise the names of each annotation is used.
}
\examples{
# Lets create a GRange object with 3 TADs and 1 metadata column:
tad.gr <- dataframes2grange(
  data.frame(
    chr = rep(1, 3),
    start = c(2e6, 3e6, 4e6),
    end = c(3e6, 4e6, 5e6),
    type = c("a", "b", "a")
  ),
  data.frame(chr = "1", size = 6e6),
  metadata.mcols = 4
)
tad.gr

# plot those TADs
# use those TADs as annotation tracks (with the metadata column to group annotations):
TADplot(
  tad.gr = tad.gr,
  chr = 1, start = 2e6, stop = 5e6,
  annot.gr = tad.gr, annot.col = 1
)
}
