% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TADshuffling.R
\name{TADshuffling}
\alias{TADshuffling}
\title{randomize domains order}
\usage{
TADshuffling(tad.gr)
}
\arguments{
\item{tad.gr}{\code{GRanges} object with domains. The chromosomes lengths must be stored as metadata (see \code{dataframes2grange()}).}
}
\value{
\code{GRanges}
}
\description{
From an ordered set of TADs, this function return un random set of TADs.
}
\details{
\code{TADshuffling} shuffle the TADs order for each chromosomes (i.e keep the TADs sizes).
The TADs input should not have gaps (region between two TADs) excepted for the extremities of chromosomes.
}
\examples{
#get domains from boundaries:
boundaries.gr = dataframes2grange(tad_HCT116_5kb.bed, human_chromsize)
domains.gr = boundary2domain(boundaries.gr)

TADshuffling(domains.gr)

}
