% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewPointInteract.R
\name{viewPointInteract}
\alias{viewPointInteract}
\title{plot view point interaction}
\usage{
viewPointInteract(
  matrix.lst,
  bin.width,
  vp.start,
  vp.stop,
  start = NULL,
  stop = NULL,
  self_interaction = FALSE,
  norm = FALSE,
  log2 = FALSE,
  colors.lst = c("#66C2A5", "#FC8D62", "#8DA0CB", "#E78AC3", "#A6D854", "#FFD92F",
    "#E5C494", "#B3B3B3")
)
}
\arguments{
\item{matrix.lst}{List of \code{dgCMatrix} or \code{matrix} object for only one chromosome.}

\item{bin.width}{Bin width of the matrix in base pair.}

\item{vp.start}{Start of the view point in base pair.}

\item{vp.stop}{Stop/end of the view point in base pair.}

\item{start, stop}{Region of interest in base pair. Default is NULL to use the entire chromosome (i.e. entire matrix).}

\item{self_interaction}{logical. Whether or not to add interactions within the view-point.}

\item{norm}{Normalized the interaction count by the average interaction of the view point (mean interaction along the matrix). Default is FALSE.}

\item{log2}{logical. Use the log2 of the matrix values. Default is \code{FALSE}. Note that if TRUE, interaction with 0 count are removed from the analysis.}
}
\value{
ggplot
}
\description{
This function (also known as virtual4C) calculates the average number of interactions of a region of interest (e.g. a TAD) with respect to the rest of the matrix.
}
\examples{
# get domains from boundaries:
boundaries.gr = dataframes2grange(tad_HCT116_5kb.bed, human_chromsize)
domains.gr = boundary2domain(boundaries.gr)

MATplot(mat_HCT116_chr19_50kb,
    start = 7e6, stop = 15e6,
    bin.width = 50e3, log2 = TRUE,
    tad.upper.tri = domains.gr["chr19_11597500",])

#plot interactions of the domain above, between 10Mb and 15Mb
viewPointInteract(matrix.lst = list(HCT116 = mat_HCT116_chr19_50kb), bin.width = 50e3,
  vp.start = 11597500, vp.stop = 12672500,
  start = 7e6, stop = 15e6,
  self_interaction = FALSE) #FALSE to not plot intra-TAD interactions




}
